requirements.txt
alembic==1.18.4
biopython==1.86
certifi==2026.2.25
charset-normalizer==3.4.5
colorlog==6.10.1
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
idna==3.11
imbalanced-learn==0.14.1
joblib==1.5.3
kiwisolver==1.4.9
lightgbm==4.6.0
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
numpy==2.4.2
optuna==4.7.0
packaging==26.0
pandas==3.0.1
pillow==12.1.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.5
SQLAlchemy==2.0.48
threadpoolctl==3.6.0
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
